# docs/智能内容运营平台/设计/设计细节/openapi.yaml
openapi: 3.0.3
info:
  title: Agentic Content Studio API
  version: 1.0.0
  description: 智能内容运营平台 API，覆盖账号、选题、内容、Agent、审核、发布和数据复盘主流程。
servers:
  - url: http://localhost:3002/api
    description: 本地 API 服务
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
    ContentIdPath:
      name: contentId
      in: path
      required: true
      schema:
        type: string
    VersionIdPath:
      name: versionId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    Platform:
      type: string
      enum: [WECHAT, XIAOHONGSHU, DOUYIN, VIDEO_CHANNEL, BILIBILI, ZHIHU, OTHER]
    ContentStatus:
      type: string
      enum: [DRAFT, PENDING_GENERATE, GENERATING, PENDING_REVIEW, REJECTED, APPROVED, PENDING_PUBLISH, PUBLISHING, PUBLISHED, FAILED, REVIEWED, ARCHIVED]
...
    ReviewStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
    PublishStatus:
      type: string
      enum: [PENDING, PUBLISHING, SUCCESS, FAILED, CANCELLED]
...
    AccountProfileUpsert:
      type: object
      properties:
        positioning:
          type: string
        targetAudience:
          type: object
          additionalProperties: true
        contentStyle:
          type: string
        titlePreference:
          type: string
        coverPreference:
          type: string
        tone:
          type: string
        forbiddenWords:
          type: array
          items:
            type: string
        contentBoundary:
          type: string
        publishStrategy:
          type: string
...
    ContentVersionPatch:
      type: object
      properties:
        accountId:
          type: string
        title:
          type: string
        body:
          type: string
        coverText:
          type: string
        tags:
          type: array
          items:
            type: string
        formatConfig:
          type: object
          additionalProperties: true
        status:
          $ref: '#/components/schemas/ContentStatus'
...
    ReviewCreate:
      type: object
      required: [contentId]
      properties:
        contentId:
          type: string
        versionId:
          type: string
    ReviewReject:
      type: object
      properties:
        comment:
          type: string
...
    PublishingTaskCreate:
      type: object
      required: [versionId, accountId]
      properties:
        versionId:
          type: string
        accountId:
          type: string
        scheduledAt:
          type: string
          format: date-time
          nullable: true
...
    AnalyticsSync:
      type: object
      required: [publishRecordId]
      properties:
        publishRecordId:
          type: string
    ReportGenerate:
      type: object
      required: [contentId]
      properties:
        contentId:
          type: string
...
    AccountCreate:
      type: object
      required: [platform, accountName]
      properties:
        platform:
          $ref: '#/components/schemas/Platform'
        accountName:
          type: string
        accountType:
          type: string
        ownerId:
          type: string
...
paths:
...
  /accounts:
    get:
      tags: [Accounts]
      summary: 平台账号列表
      parameters:
        - name: platform
          in: query
          schema:
            $ref: '#/components/schemas/Platform'
        - name: authStatus
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
    post:
      tags: [Accounts]
      summary: 手动创建平台账号
      x-implementation-status: planned
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreate'
      responses:
        '200':
          description: OK
  /accounts/{id}:
    get:
      tags: [Accounts]
      summary: 平台账号详情
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
    patch:
      tags: [Accounts]
      summary: 更新平台账号
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                platform:
                  $ref: '#/components/schemas/Platform'
                accountName:
                  type: string
                accountType:
                  type: string
                authStatus:
                  type: string
                ownerId:
                  type: string
                rawData:
                  type: object
                  additionalProperties: true
      responses:
        '200':
          description: OK
...
  /account-profiles:
    get:
      tags: [AccountProfiles]
      summary: 按账号查询画像
      parameters:
        - name: accountId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /account-profiles/{accountId}:
    put:
      tags: [AccountProfiles]
      summary: 创建或更新账号画像
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountProfileUpsert'
      responses:
        '200':
          description: OK
...
  /versions/{versionId}:
    get:
      tags: [Versions]
      summary: 版本详情
      parameters:
        - $ref: '#/components/parameters/VersionIdPath'
      responses:
        '200':
          description: OK
    patch:
      tags: [Versions]
      summary: 更新平台版本
      parameters:
        - $ref: '#/components/parameters/VersionIdPath'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentVersionPatch'
      responses:
        '200':
          description: OK
...
  /reviews:
    get:
      tags: [Reviews]
      summary: 审核列表
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ReviewStatus'
      responses:
        '200':
          description: OK
    post:
      tags: [Reviews]
      summary: 提交审核
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCreate'
      responses:
        '200':
          description: OK
  /reviews/{id}:
    get:
      tags: [Reviews]
      summary: 审核详情
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
  /reviews/{id}/approve:
    post:
      tags: [Reviews]
      summary: 审核通过
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
  /reviews/{id}/reject:
    post:
      tags: [Reviews]
      summary: 审核驳回
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewReject'
      responses:
        '200':
          description: OK
  /publishing/tasks:
    get:
      tags: [Publishing]
      summary: 发布任务列表
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/PublishStatus'
      responses:
        '200':
          description: OK
    post:
      tags: [Publishing]
      summary: 创建发布任务
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishingTaskCreate'
      responses:
        '200':
          description: OK
  /publishing/tasks/{id}:
    get:
      tags: [Publishing]
      summary: 发布任务详情
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
  /publishing/tasks/{id}/publish:
    post:
      tags: [Publishing]
      summary: 执行发布
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
  /publishing/tasks/{id}/cancel:
    post:
      tags: [Publishing]
      summary: 取消发布
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
  /analytics/sync:
    post:
      tags: [Analytics]
      summary: 同步发布数据
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyticsSync'
      responses:
        '200':
          description: OK
  /analytics/contents/{contentId}:
    get:
      tags: [Analytics]
      summary: 内容数据指标
      parameters:
        - $ref: '#/components/parameters/ContentIdPath'
      responses:
        '200':
          description: OK
  /analytics/reports:
    get:
      tags: [Analytics]
      summary: 复盘报告列表
      parameters:
        - name: contentId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
  /analytics/reports/generate:
    post:
      tags: [Analytics]
      summary: 生成复盘报告
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportGenerate'
      responses:
        '200':
          description: OK
  /analytics/reports/{id}:
    get:
      tags: [Analytics]
      summary: 复盘报告详情
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
...